requirements.txt
python-telegram-bot[job-queue]==22.7
flask==3.1.0
reportlab==4.4.1
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
